created modules.py for all database query functions
